[ffx][ota] Allow specifying ota server port
This adds `--product-bundle-port`  option to `ffx target update` commands. This is the port
to start the OTA repo server on when using `--product_bundle`.

It can be configured be setting `repository.ota_port`. The default is 0,
which indicates a random unassigned port.

This is added to support using a specific port with other tools to
support remote workflows.

Bug: 388844218
Change-Id: Ie21f388892b2a36adca32d8d2a07a28c7a0686e3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1190192
Reviewed-by: Amit Uttamchandani <amituttam@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>