[fuchsia-controller] Disallow undefined symbols in C extensions
This change adds an action that generates a linker script to
create a weak symbol for each dynamic symbol in the python interpreter
prebuilt that's part of the Python C API found. By doing so, a shared
library for a Python C extension can be built while disallowing
undefined symbols.

This likely all sounds a bit obtuse. After all, it's generally okay to
have undefined symbols dynamically linked to symbols in other shared
libraries. The problem is that the Fuchsia tree does not contain
libpython and so we cannot currently build the shared library in such a
way that the linker can dynamically link the Python C API symbols to
libpython.

Fixed: 295680434
Change-Id: I78a6efde9964ef0f7f237948c719ab6b9dd77c29
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1178099
Reviewed-by: Steven Grady <slgrady@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Charles Celerier <chcl@google.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>