[kernel][vm] Make VmObject::user_id_ a relaxed atomic
Use a RelaxedAtomic to store the user_id_ instead of guarding with the
lock. The user_id_ is already const-ish and the lock is only preventing
read/write tearing with debug code, which can be achieved with an
atomic.

The parent_user_id_ for physical vmos is changed to be true const and
initialized in the constructor.

Not needing to acquire the lock of the parent when querying the
parent_user_id simplifies the locking semantics for fine grained VMO
locking.

Bug: 338300943
Change-Id: I5c2d29df6b6f5f308c7285a35b178507117ccfd6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1187434
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Sage Barreda <sagebarreda@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>