[fxfs] Ensure that link accesses a consistent file.
Adds a loop since we need to take a lock on the file object but we don't
know what the object is until we look it up. Ensure under lock that the
file in that location hasn't changed. Otherwise it may have been
unlinked between the lookup and the lock, and may be marked for purge.

As long as the file is still there after the lookup under lock, then it
can't be marked for purge, and whatever other races may have occurred
are irrelevant since this file was in the directory under that name at
some pointed during the call and the thing found there is still valid to
be linked.

Adds a test to catch the naive TOCTOU based on the requisite transaction
locks for the operations.

Updates a NOT_FOUND error to be INCONSISTENT, since the code path there
is a result of looking up an object id that no longer exists.

Bug: b/388149864
Change-Id: I1e0e37865e05ad818e29ec6a1c2f13095f964be9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1190232
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>