[tee] Implement the transient storage interface
Well, most of it. This change introduces the key types associated with
transient objects, but with stubs to be properly implemented when we add
support for particular cryptographic schemes. The transient storage
interface is in turn fully implemented in terms of these types.

Bug: 371213067
Change-Id: I8c219dbe32529f380086bcf344e27e5fe1b86ac0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1183052
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>