[virtual_audio] Simplify device shutdown
Virtual audio devices will be given a callback to be invoked when they
have completely shutdown. A device is completely shutdown when its fidl
server and driver have completely shutdown. These drivers will also be
given a callback that they will invoke only when the driver is
completely shutdown. When a driver is completely shutdown is determined
by the different implementations of the driver.

The virtual_audio driver itself will keep track of which devices have
been shutdown and remove them. Originally, devices that shutdown were
not removed and instead marked as unbound.

This simplifies the logic for waiting for all devices to shutdown. This
also reduces the need for a cyclic dependency between
`VirtualAudioDevice` and `VirtualAudioDriver`.

Bug: 388880875
Change-Id: I1b3ba530fc0e57808dd804add071be97600927e5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1189216
Commit-Queue: Nick Eaton <nickeaton@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>
Reviewed-by: Martin Puryear <mpuryear@google.com>