[ui] rename ActivityManager as InteractionStateHandler
This change renames ActivityManager as InteractionStateHandler, to
correspond to the service's new role as an InputHandler registered with
the Input Pipeline. There are also changes to documentation to help
disambiguate the class from broader notions of "activity" in the fuchsia
system, such as in Power Framework, where there are concepts like the
System Activity Governor (and something already called "ActivityHandler").

Bug: 389664333
Test: fx test interaction-state-test
Test: fx test input_pipeline_lib_tests
Change-Id: Idab8b01e3ec483ce80e18c875ed3e134f529fd1a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1187714
Reviewed-by: Alice Neels <neelsa@google.com>
Commit-Queue: Caroline Liu <carolineliu@google.com>