[auth] Unset GOOGLE_APPLICATION_CREDENTIALS
Hide GOOGLE_APPLICATION_CREDENTIALS environment variable
from bazel when it is configured to use gcert-authentication.
This addresses the issue where bazel looks for this credential
file when it is not needed and fails ungracefully.

We still have to keep this variable for the exceptional
uses of remotetool, which doesn't have gcert-auth support
and needs the old gcloud method.
remotetool is used for on-demand downloading with
rbe_mode="low_bandwidth_remote".

Bug: 358163278
Change-Id: I5aaf7584590d3d566dd61d53265d58d790b271c3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1190203
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Fang <fangism@google.com>
Reviewed-by: David Turner <digit@google.com>