[devicetree] Introduce "status" property.
StatusProperty reflects the property named "status" as described in
devicetree documentations (Chapter 2.3.4).

https://devicetree-specification.readthedocs.io/en/latest/chapter2-devicetree-basics.html#status

This chain intends to add support to skipping not "okay" nodes. While
the spec only refers to "okay" for ok status, other operative systems
support "ok" as well.

Additionally we cache this property so it's cheap for multiple matchers
to check the state of the node, which ended up renaming CellCounts to
CachedProperties.

Bug: 385360435
Change-Id: Ida62bf91ad94328c17470be0fb555f943f787851
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1191093
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>