[block_server][rust] Enable trace flows
Adds trace flow steps for the start and end of transactions, aligned
with the tracing available in the C++ block server.

I also changed the rust block client to attempt to reuse the request's
trace_flow_id so we can link the entire flow end-to-end.  Unfortunately,
a bit more work will be required here, since there's no way to specify
the trace flow ID to the block client.  I'll do that in a follow-up CL.
(For now, the behaviour is that the block client makes a new trace flow,
so there is a disjoint flow nested within the other flow, in processes
that implement block_server by forwarding to block_client (e.g. the GPT
component).

I also removed the future trace in
block_client::RemoteBlockClient::send, because it is redundant with the
flows.

Bug: 375083001
Change-Id: Ia61fb312c382412d075f15c7590fab450dd4682f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1191133
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>