[starnix] Fixing flaky suspend test
Test flaked due to OnSignals::new only polling after being registered.
By adding a poll to the ASLEEP signal before calling suspend_context,
we are able to guarantee that the signal is registered in advance of the
function running so that a race condition is avoided.

Fixed: 383362828

Change-Id: I0a3c56cea327adb513c47e7215f10a3dc3544ff4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1190112
Commit-Queue: Olivia Nordquist <nolivia@google.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Vickie Cheng <vickiecheng@google.com>