[zxtest] Invert conditions in streams.
This allows us to generate if/else branches that are always paired,
avoiding unexpected results when a single if branch is expanded.

if(true)      if (true)
   MACRO   ->     MACRO- if (false)
else       ->            else

To:
if (true)
    MACRO- if
    MACRO- else
else

This also uncovered an exisitng bug in a macro
LIB_ZXTEST_HEXDUMP_PRINTER and introduced a warning dangling-else.

The first one, is due to the macro parameter being named size, and the
variable used in the macro body being byte_count. Just happens that it
matches the context where it is used. This has been fixed. Some
indirection is needed, to generate storage for the variable when the
value comes from expression such as sizeof(T).

The second one, needs some specifc tricks to trick the compiler into not
identifying the dangling else.

Fixed: 388300636
Change-Id: I3873b34e92c7206cba251c5616f42cab160198fe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1191212
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>