[fuchsia.io] Add flags for unnamed temporary files
We want to add support to open an unnamed temporary files. This file can
be linked into the filesystem to make it persistent. However, it can
also be created to NOT be linkable.

This defines a new member of Open3's `Flags` to indicate to open the
object as a temporary unnamed object. A protocol is expected to be
specified with this flag to indicate the type of object to be created.
Currently we will only support `PROTOCOL_FLAGS`.

If we want to create an unnamed temporary file that is not linkable,
this is indicated by setting the flags `FLAG_CREATE_AS_UNNAMED_TEMPORARY
| FLAG_TEMPORARY_AS_NOT_LINKABLE`.

Bug: 293950309
Change-Id: I394907ed85c17bfa2f71bcf6bc93948bf8017663
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1185112
API-Review: Chris Suter <csuter@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Rui Qi Sim <rqsim@google.com>