[uart] Introduce KernelDriver/Driver::mmio_range
Introduce helper struct uart::MmioRange, which holds a tuple
reprensenting the memory range in the addressing provided by it's
configuration.

Introduce concept uart::MmioDriver<T> which is slightly
less verbose than `std::is_same_v<....., zbi_dcfg...>`
which evaluates the Driver::IoType(Pio, Mmio8, Mmio32) and decides
whether it is MMIO or not.

When a type complies with the MmioDriver concept, then uart::Driver and
uart::KernelDriver will provide additional accessors for obtaining the
device's MMIO range as 'uart::MMioRange mmio_range()'.

`io_slots` are adjusted to determine the proper length of the range.
The provided range is NOT required to be page aligned and it is
the callers responsibility to align to meet their requirements.

Bug: 42084617
Change-Id: I522e2ff79d00b7d3c2990824db4b957ae8e4c77d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121694
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>