[uart] Introduce TakeUart.
For both all::KernelDriver and uart::KernelDriver TakeUart moves the
uart from one object to another.

For uart::all::Driver we introduce std::monostate as the move_from
state, and attempting to visit the uart in this state, will cause an
abort. This is only useful to detect logging in certain spots where the
handoff is in progress and it's for a very short window.

In order to simplify the dispatch and handling of this state, an
internal helper class, VariantVisitor is introduced. It essentially
wraps the provided callback provided to visit and special cases
std::monostate to produce an abort if it's the active instance, otherwise
forwards the arguments to the provided callback.

That said, uart() accessors cannot be removed for out of tree
dependencies at this point. As soon as those are migrated, we can follow
up and remove them from here.

Bug: 42084617
Change-Id: Ib1f49e1f4b05b33c2ee4c6f6f129e3474b1fe3a1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121695
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>