[fxfs] Reduce first superblock extent to 4kB.
Because we expect at least 64kB of free space, this necessarily leads
to two 64kB allocations immediately after FS creation. Prior truncate
logic had a two bugs that this also addresses:
  1. Assumed file was minimally sized in writer code.
  2. Truncate of superblock after writing didn't take into account
     the 64kB free space requirement, so when a second extent is
     needed just to provide this free space, this led to allocate/free
     cycles for every superblock write.

There is also a bug fixed in make_blob_image due to insufficient free space in the target image and a false assumption in an fallocate test that space will not be freed in the regular course of applying transactions.

Change-Id: I1aa82a075aae3d4f00b4cb8892f8eb08619bc39e
Bug: b/381277474
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1189932
Commit-Queue: Aaron Drew <ripper@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>