[fxfs] clean up checksum chunking in multi_overwrite
The range math that broke up the vector of checksums that came back from
write was broken in a couple ways. First, it was calculating the device
range incorrectly because it was using a value that had an offset
included in it. Second, it wouldn't push the last chunk because it only
pushed when the bitmap value flipped.

This simplifies the logic so it's much more readable and understandable.
There are probably optimizations but it's not worth the obfuscation
here.

We also add a couple checks to the omni-test for multi_overwrite that
makes sure the checksums it puts into the transaction make sense to
prevent regressions here.

Fixed: 383564671
Change-Id: I9e99ecc0d4feead18ce171cf8419f494f30b796e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1189239
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>