[shutdown-shim] Inspect for Collaborative Reboot
Integrate shutdown-shim with the Inspect framework, and add counters for
the currently scheduled collaborative reboots.

Because shutdown-shim is a built-in component of Component Manager, it
uses a local instance of the `Inspector` type, as opposed to the
globally-scoped `Inspector` (which is used by Component Manager). This
ensures that Component Manager and shutdown-shim each have their own
distinct Inspect sections.

Bug: 380873116
Change-Id: Icb6ecb445fa89c0019c25f64433f4268db08f95d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1186272
Reviewed-by: Amanda Tait <atait@google.com>
Fuchsia-Auto-Submit: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Hai Bi <bihai@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>