[storage-benchmarks] Use config capabilities
Rather than sniffing for the system configuration, read config
capabilities (fuchsia.fshost.*) to detect how to run the storage
benchmarks.

To achieve this, I added a root shard which routes some capabilities to
the core realm, which is bundled into the "testing_support" AIB.  These
routes aren't particularly sensitive, but I thought it would be better
to start a precedent of keeping capabilities that are exclusively routed
for testing, in shards which are bundled with testing support.

This change also allows the benchmarks to run on storage-host builds
with Fxblob and fixes a few minor issues.

Bug: 372555079
Change-Id: I9f945c7f8cc86d80fa0d401b27e34980f76f5f2c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1190816
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>