[idk][build] Remove abstractions around vulkan_layers
The IDK's dependency on "//src/lib/vulkan:vulkan_layers" was wrapped in
two layers of molecules that did not contain anything else. Replace the
dependency on these molecules with a direct dependency on vulkan_layers
for clarity and to eliminate the implication that loadable modules are
special and the logical complexity of the
"fuchsia_idk_atoms_independent_of_api_level_for_one_cpu_arch" target.

Bug: 347094435
Change-Id: I97bf0b67a7e6121ed5c67f1289145926f33db5c9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1191196
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>