[bazel_sdk] avoid loading fuchsia_sdk in fuchsia_clang
This CL avoids loading the fuchsia_sdk repository inside of the
fuchsia_clang repository by hardcoding the list of supported CPUs. This
is safe because the CPUs will always exist in the official SDK. For
custom builds and in-tree builds where the CPUs might not exist it is
still safe as long as someone doesn't try to target the missing CPU.
This is guaranteed to be the case when building in-tree due to how the
build is structured and when using a custom CPU we document the
constraint.

Bug: 369638604
Change-Id: Ice5cbdc8d5151d9076e5886d60a22a1ebd9ba1d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1191925
Reviewed-by: Darren Chan <chandarren@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>