[multifuchsia] Touch args.gn upon checkout and rebase
This ensures that upon the first `fx build` in the new checkout or
rebased directory an `fx gen` is run to ensure the gn build graph is
correct. This works around an edge case that results in ninja ending up
in a loop.

Fixed: b/391278102
Change-Id: I73e214a0eb6a7564d3060590846d1223815c3298
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1190733
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Thomas Zander <thomasalpinus@google.com>