[virtual_audio] Remove thread sync for virtual audio drivers
Remove the thread synchronization mechanisms used on
`VirtualAudioDriver`. The virtual audio driver's dispatcher is
synchronous and can be treated as a single-threaded dispatcher and so
thread synchronization is not needed.

Make the implementation of `VirtualAudioDriver` decide which dispatcher
to use in order to perform the fuchsia.virtualaudio.Device getters and
setters.

Bug: 388880875
Change-Id: I19ad88215939588b5ae6f8e93e2d170374700cad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1191032
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>