[assembly] Move serial core driver to base from bootfs
The serial driver is now a boot driver which is not the ideal case. The
reason is that it should act as the fallback driver for uart child node,
which means it will bind to uart child node when no other driver binds
to the node, as a boot driver it'll seize the node from all the base
drivers because starts earlier during device boot. This CL moves serial
driver to base driver.

Change-Id: Ib6c903f16e33d9d5e6d05ca07d6675f159f6851c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1177132
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Size-Review: James Sullivan <jfsulliv@google.com>
Commit-Queue: Zhiyi Chen <zhiyichen@google.com>