[utest][core] Remove min_iterations in DecommitRace
The DecommitRace previously asserted that we could run 100 iterations
of the pin/unpin loop within 10 seconds. This led to spurious
failures in emulated environments where the host was overloaded.

Unfortunately, solving this CL isn't as simple as removing the 10
second time limit. The code was originally structured to just run
a fixed number of iterations of the loop, but this frequently caused
the test to hit infrastructure timeouts on RISC-V builders. The 10
second time limit was introduced to prevent those flakes, and is
thus necessary.

Instead, this CL just removes the minimum assertion check. Note
that this does decrease the likelihood of inducing the race this
test was originally intended to solve, but it's the best we can
do given limited CPU time.

Fixed: 392093417
Change-Id: Ic46069c2afccec3379e156f9a8899903f126bb57
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193120
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>