[uart] Introduce Rx/TxInterrupt to properly annotate methods with TA
Instead of passing arbitrary lambdas we are passing an object wrapper
that hold the lambdas and other objects. It provides the TA annotations
to enforce holding certain capabilities when calling such objects.

The API is a compile-time contract that provides methods to access the
capability disable the IRQ (TX or RX).

Change-Id: I848cb46d73a3307fb97c70901a2ca9388b985ecc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1119761
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>