[graphics][msd-vsi-vip] Add GtestPrinter
Add a GtestPrinter that will log the Gtest output to the driver's
log. This makes it easier to see test output with

 ffx log --filter msd_vsi_test

The main downside to this is maintaining the printing code, and
the extra padding that is logged to the left, but this can be
somewhat mitigated with the --hide-tags option.

Change-Id: I087bd424e8b0ec690746a373d18230635fabe355
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193116
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Craig Stout <cstout@google.com>