[ffx][target_holders] Create TargetProxyHolder type
This is a newtype around TargetProxy which allows implementation of
TryFromEnv outside of the fho crate. This is the first "Holder" type, so
it has some duplication with the fho crate, but that will be
incrementally move out of fho in order to eliminate the dependencies
which are related to device interactions so they do not negatively
impact build time of the subtools that do not require devices.

Bug: 391610009
Change-Id: Ib4e3fe6f83ee608e2bfd543274b4596e81126d7e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193235
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>