[drivers][power] Acquisition flips state flag
When we acquire a wake lease flip the flag tracking whether the
system is suspended. If the system was resumed when the
`fdf_power::WakeLease` was created, the object won't know the
system state until it changes. This might result in a lot of
unnecessary wake lease acquisitions.

MULTIPLY: aml-uart-test

Change-Id: I8a17190f84fe781857671c34472530cf55f3da33
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1192485
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Justin Mattson <jmatt@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>