[ffx][build] Remove default sdk_category from ffx_plugin()
Require that `ffx_plugin()` instances in an SDK specify the
`sdk_category` rather than defaulting to "partner". Instances that do
not specify a category are for internal use only.

Use "not-yet-specified" for the 200+ existing plugins that do not
specify a category to avoid needing to determine the correct category
for all of them in this change. Remove the category from the few
instances that previously specified "internal".

Specifying the appropriate category will allow more accurate application
of the SDK markers and related category enforcements.

Bug: 314822328
Change-Id: Ie2285927dc198b8bf087bb3328b7f9d04d5c68bd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1191537
Commit-Queue: David Dorwin <ddorwin@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Fuchsia-Auto-Submit: David Dorwin <ddorwin@google.com>