[intel-display][hdmi] Read E-EDID when a display is available.
Previously, when an HDMI display was available, the intel-display
driver sent the DDC I2C interface to the display coordinator,
and the coordinator sent I2C commands to read the extended EDID
bytes.

With this change, HdmiDisplay reads the extended EDID bytes
when a display is available, and sends the bytes to the coordinator.

Test: core.x64 booted on NUC11TNHv5 with MIMO HDMI display.
Test: core.x64 booted on NUC11TNHv5 with HDMI headless plug.
Bug: 390552176
Change-Id: Ia66f4bb90fc76949dbad9e97b110a7f445391ead
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1191929
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Yilong Li <liyl@google.com>