[ffx][target_holders] Create TargetInfoHolder type
This is a newtype around TargetInfo which allows implementation of
TryFromEnv outside of the fho crate.

This is part of eliminating the dependencies in fho
which are related to device interactions so they do not negatively
impact build time of the subtools that do not require devices.

Bug: 392174744
Change-Id: I54b0a9cb747368559f0a1a0fb96d17aa508d1956
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193395
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>