[component] Use r* for OpenServiceRoot
As part of enforcing hierarchal rights with Open3, we need to ensure the
set of readable directory rights is included whenever opening /svc. This
ensures the availability of many expected rights on the resulting
connection (e.g. enumerating services/protocols, or calling stat on the
directory).

This brings the implementation in line with //sdk/lib/sys which was
updated in I1940c0a0f631347a35231a122555e56c07892fa7.

Bug: 346585458
Change-Id: I6696bc4c383f51427ea175b9a9fb5f40911b9368
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1190772
Commit-Queue: Brandon Castellano <bcastell@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>