[fshost] reenable unformatted_small_disk on minfs
This test was disabled "temporarily" a while ago because of a flake bug,
but it was never enabled again. That flake bug is long closed, and some
sleuthing indicates whatever underlying issue that this was disabled for
was fixed.

Change-Id: I6adbc89059734fb1e8e23f382bb18b3a432c0a39
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1192774
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>