[zircon] Try to dump secondary CPU state if not online
This change wires up lockup_internal::DumpRegistersAndBacktrace() to the
error handling around secondary CPUs not coming online. If we have such
a facility, then we dump that debugging feedback and can panic in good
conscience; otherwise, we carry on with plain OOPSes, as we do today.

Bug: 389702612
Change-Id: Ie0d518abb4e3cc5f938a31fd84babed1f1335181
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1191192
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>