[ffx][fho] Refactoring DeviceLookupDefaultImpl
This refactors DeviceLookupDefaultImpl out of fho and into
target_holders, where it is primarily used when implementing
TryFromEnv for target related objects.

This is part of the project to eliminate the dependencies of fho
which are related to device interactions so they do not negatively
impact build time of the subtools that do not require devices.

Bug: 392318888
Change-Id: Ib0a5a86e6cdcb5b9ca9f3af39772177ad62b3d5c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193580
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>