[ffx][target_holders] Create DaemonProxyHolder type
This is a newtype around DaemonProxy which allows implementation of
TryFromEnv outside of the fho crate.

This also refactors the daemon_protocol() method used to initialize
protocol proxies from the daemon connection.

This is part of eliminating the dependencies in fho which are related to
device interactions so they do not negatively impact build time of the
subtools that do not require devices.

Bug: 392173931
Bug: 392174459
Change-Id: Iff7036c248bbd1366ed7b262246dbd076d3b3440
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193587
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>