[usb-overnet][test] Get tests compiling again
Updates to new test framework for drivers, while still using zxtest
because of the banjo mocks using it extensively. This test was disabled
while it was still a dfv1 driver, so a lot has changed since then and
the tests had to be updated to match dfv2 ways of doing things.

Bug: 382314045
Multiply: overnet-usb-test
Change-Id: I61c5e7e53e71597eb739cf0a109b344a0c69c53a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1172129
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>