[ffx][fho] Refactoring toolbox() and moniker() out of fho
This refactors moniker() and toolbox() out of fho and into
target_holders, where it is primarily used when implementing
TryFromEnv for target related objects.

This is part of the project to eliminate the dependencies of fho
which are related to device interactions so they do not negatively
impact build time of the subtools that do not require devices.

Bug: 392173451
Bug: 392174212
Change-Id: Id7a3375f12602ffd504f922ff95513fae06c3d79
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193571
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>