[kernel][phys][arm64] Hand off address space(s)
As already done with riscv64 and x86, this is a major step toward
physboot properly creating the kernel's address space: it hands off the
address space as it is currently constructed, upper and lower - along
with the kernel's mapping with its final page protections in place -
enabling the kernel to bootstrap from the already-constructed identity
map.

This is the second reland of fxrev.dev/1173133; the reverts were due to
instability around secondary boots. It is now believed that this was
due to stores of boot info on the boot CPU - as well as instruction
memory itself - not being flushed in time for loads/execution on
secondaries. This new version addresses that.

Bug: 42164859, 389702612
Change-Id: I2420364770f652568f7f86efb2fbc80a68684c20
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1190177
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>