[ffx][target_holders] Create RemoteControlProxyHolder type
This is a newtype around RemoteControlProxy which allows implementation of
TryFromEnv outside of the fho crate.

This is part of eliminating the dependencies in fho which are related to
device interactions so they do not negatively impact build time of the
subtools that do not require devices.

Bug: 392174888
Change-Id: Iff6a6400626eefd653eb4ba0a35555a41f7c2c7f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1193755
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>