[fuchsia.io] Replace GetFlags/SetFlags at NEXT
Rename existing GetFlags and SetFlags methods to DeprecatedGetFlags and
DeprecatedSetFlags. Replace with GetFlags2/SetFlags2 by renaming them
back to GetFlags/SetFlags.

This CL does not remove or change any existing usages of either method.
Transitioning callers to the new method will be handled as a follow up.

zxio and fdio both gracefully fallback to the old method if the server
does not have support for the new ones, however this may incur a latency
penalty. Once we complete out-of-tree server support, we will transition
all callers off of the deprecated method and mark it as removed.

Unfortunately this means we'll incur a temporary size increase while we
need to support both methods, but this should be resolved as soon as we
complete out-of-tree support.

Bug: 376509077
Test: No functional change, rename only.
Change-Id: If7c70b0bac6febf572e81d9da0bb99cdedd9eafd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1189354
API-Review: Gary Bressler <geb@google.com>
Reviewed-by: David Dorwin <ddorwin@google.com>
Reviewed-by: Rui Qi Sim <rqsim@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>
Reviewed-by: Gary Bressler <geb@google.com>