[drivers] Use Open instead of Connect for dev-topological
As part of the ongoing fuchsia.io migration, there are new rights in
place for certain operations like getting attributes and enumerating
directory entries. This means that certain cases where opening a
directory as a service will fail, as we do not request any rights when
connecting to a service.

These rights are implicitly added in Open1 but are not added implicitly
in Open3. To avoid issues with switching fdio over to Open3, we ensure
that dev-topological is always opened as a directory with the correct
rights to directory watching can operate correctly.

In some cases it was easier to operate directly on the test realm's
exposed directory, so this was added where possible. In other cases,
we open the corresponding directory with Open3 instead of using Connect.
Also add watcher function for unowned client ends.

Bug: 324111518
Change-Id: I7f1c3dc4ea93a766ef323dd1bc7800be92bdd9b5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1165732
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Martin Puryear <mpuryear@google.com>
Reviewed-by: Jianpeng Chao <chaopeng@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>