[ffx] Fix expiration of "ad-hoc" targets
If you make an ad-hoc connection (e.g. "ffx -t 1.2.3.4 target echo"), we
will add 1.2.3.4 to the daemon's target collection, but due to incorrect
logic, it gets an "expiration", which means that if we don't see any
mDNS packets from that target (and there's no reason to expect any --
it wasn't "discovered"), then we eventually (after 25 seconds) expire
the address.

After which, the next request of "ffx -t 1.2.3.4 target echo" would
not find the target by its address, and it would establish a second
host-pipe connection. Then the daemon would wait for Overnet to
inform it of the new connection, so it could try to establish an RCS
connection. But from Overnet's point of view, there is no new target;
it already knew about it.

So the daemon never completes the RCS connection, resulting in a
time-out in the client.

The fix is simply to call it a "manual" target, with no expiry.

Bug: 392682202
Change-Id: I5bb02fdf27e42fa9c6780783ecba9e5937f9fb96
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1194236
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>