[zircon][vm] Externalize freeing pages from ReleaseOwnedPagesLocked
Instead of freeing the pages inside ReleaseOwnedPagesLocked, we now
return a list of pages that need to be freed to the callers. This CL
introduces a RAII object to ensure that the callers actually free the
pages.

Bug: b/338300943
Change-Id: I09789756a9e97d826a95277f55cf934fc1398df3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1192355
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>