[fdio] Don't use fdio_service_connect for opening /svc
The /svc should be opened as a directory and not a service. In most
cases, opening the /svc directory can be done with the component helper
library, or by using `fdio_service_connect` with the full path to the
service, including the /svc prefix.

Bug: 324111518
Change-Id: I7943b7b81a6048985027503c5a7d127109559c1c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1165733
Reviewed-by: David Pursell <dpursell@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>