[dl] Assert module_ in SetAbiName
SetAbiName is a no-op if `module_` isn't set on the `DecodedModule`
instance. A caller may set the name for the module before setting the
abi module and will not know that the name setting does not stick.

This CL adds an assert that `module_` is set in `SetAbiName` so callers
are enforced to set the module before its name.

Bug: 338238626
Cq-Include-Trybots: luci.turquoise.global.try:run-postsubmit-tryjobs
Change-Id: Idc79e17efc27d048250a63a1cfb4945f0454766b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1192132
Fuchsia-Auto-Submit: Caslyn Tonelli <caslyn@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>